mv walkDir to utils
