+++ packages/markdowndb/src/lib/utils/index.ts
@@ -1,0 +1,1 @@
+export { default as recursiveWalkDir } from "./recursiveWalkDir";
